Add ntp and make sure time is correct
Chef was not running initially because the time on the guest was
too far off. Whether or not UTC was checked in VirtualBox, the guest
time was being set as if the EDT time of the host was actually UTC
time, so just using ntp so the user won't have to worry about it.